+++ src/renderer/src/main.tsx
@@ -8,6 +8,7 @@ import {
 import { DEFAULT_OPEN_NATIVE_CONSOLE } from "@shared/open-native-console";
 import { DEFAULT_UI_DENSITY } from "@shared/ui-density";
 import "@mantine/core/styles.layer.css";
+import "@mantine/dates/styles.layer.css";
 import "@mantine/carousel/styles.css";
 import "@mantine/notifications/styles.layer.css";
 import "@mantine/spotlight/styles.layer.css";